s3-backup-system/requirements.txt
aws-cdk-lib==2.149.0
constructs==10.3.0
boto3
